testing and defined for
