+++ policyengine_us/tests/policy/baseline/gov/irs/income/taxable_income/deductions/overtime_income/overtime_income_deduction.yaml
@@ -1,7 +1,7 @@
 - name: Ineligible cardholder present
   period: 2025
   input:
-    overtime_income_deduction_ineligible_cardholder_present: true
+    overtime_income_deduction_ssn_requirement_met: false
     fsla_overtime_premium: 10_000
     adjusted_gross_income: 100_000
     filing_status: JOINT
@@ -11,7 +11,7 @@
 - name: AGI below phase out start
   period: 2025
   input:
-    overtime_income_deduction_ineligible_cardholder_present: false
+    overtime_income_deduction_ssn_requirement_met: true
     fsla_overtime_premium: 10_000
     adjusted_gross_income: 100_000
     filing_status: JOINT
@@ -21,7 +21,7 @@
 - name: Fully phased out
   period: 2025
   input:
-    overtime_income_deduction_ineligible_cardholder_present: false
+    overtime_income_deduction_ssn_requirement_met: true
     fsla_overtime_premium: 10_000
     adjusted_gross_income: 1_000_000
     filing_status: JOINT
@@ -31,7 +31,7 @@
 - name: Partially phased out
   period: 2025
   input:
-    overtime_income_deduction_ineligible_cardholder_present: false
+    overtime_income_deduction_ssn_requirement_met: true
     fsla_overtime_premium: 10_000
     adjusted_gross_income: 190_000
     filing_status: JOINT
