+++ policyengine_us/tests/policy/baseline/gov/irs/income/taxable_income/deductions/overtime_income/overtime_income_deduction_ssn_requirement_met.yaml
@@ -8,13 +8,13 @@
         is_tax_unit_spouse: false
       person2:
         ssn_card_type: OTHER_NON_CITIZEN
-        is_tax_unit_head: true
-        is_tax_unit_spouse: false
+        is_tax_unit_head: false
+        is_tax_unit_spouse: true
     tax_units:
       tax_unit:
         members: [person1, person2]
   output:
-    overtime_income_deduction_ineligible_cardholder_present: true
+    overtime_income_deduction_ssn_requirement_met: false
 
 - name: Ineligible cardholder present, but not head or spouse
   period: 2025
@@ -32,7 +32,7 @@
       tax_unit:
         members: [person1, person2]
   output:
-    overtime_income_deduction_ineligible_cardholder_present: false
+    overtime_income_deduction_ssn_requirement_met: true
 
 - name: Ineligible cardholder not present
   period: 2025
@@ -50,4 +50,4 @@
       tax_unit:
         members: [person1, person2]
   output:
-    overtime_income_deduction_ineligible_cardholder_present: false
+    overtime_income_deduction_ssn_requirement_met: true
