+++ policyengine_us/variables/gov/irs/income/taxable_income/deductions/tip_income/tip_income_deduction.py
@@ -7,6 +7,7 @@ class tip_income_deduction(Variable):
     label = "Tip income deduction"
     unit = USD
     definition_period = YEAR
+    defined_for = "tip_income_deduction_ssn_requirement_met"
 
     def formula(tax_unit, period, parameters):
         person = tax_unit.members
@@ -19,8 +20,4 @@ class tip_income_deduction(Variable):
         phase_out_amount = agi_excess * p.phase_out.rate
         total_tip_income = tax_unit.sum(tip_income)
         capped_tip_income = min_(p.cap, total_tip_income)
-        phased_out_tip_income = max_(0, capped_tip_income - phase_out_amount)
-        ineligible_cardholder_present = tax_unit(
-            "tip_income_deduction_ineligible_cardholder_present", period
-        )
-        return phased_out_tip_income * ~ineligible_cardholder_present
+        return max_(0, capped_tip_income - phase_out_amount)
